@@ -4,6 +4,7 @@ from decimal import Decimal
 from os.path import abspath
 from tempfile import gettempdir
 
+import django
 from django.conf import settings
 from django.core.validators import (
     validate_ipv4_address,
@@ -18,7 +19,7 @@ import pytest
 from model_bakery import baker
 from model_bakery.content_types import BAKER_CONTENTTYPES
 from model_bakery.gis import BAKER_GIS
-from model_bakery.random_gen import MAX_LENGTH, gen_related
+from model_bakery.random_gen import MAX_LENGTH, gen_from_choices, gen_related
 from tests.generic import generators, models
 
 try:
@@ -72,15 +73,41 @@ def custom_cfg():
 
 class TestFillingFromChoice:
     def test_if_gender_is_populated_from_choices(self, person):
-        from tests.generic.models import GENDER_CHOICES
+        from tests.generic.models import Gender
 
-        assert person.gender in (x[0] for x in GENDER_CHOICES)
+        assert person.gender in Gender.values
 
     def test_if_occupation_populated_from_choices(self, person):
         from tests.generic.models import OCCUPATION_CHOICES
 
         occupations = [item[0] for lst in OCCUPATION_CHOICES for item in lst[1]]
         assert person.occupation in occupations
+
+
+class TestGenFromChoices:
+    def test_excludes_blank_when_not_blankable(self):
+        choices = [("", "empty"), ("A", "a"), ("B", "b")]
+        gen = gen_from_choices(choices, nullable=True, blankable=False)
+        for _ in range(100):
+            assert gen() != ""
+
+    def test_excludes_none_when_not_nullable(self):
+        choices = [(None, "none"), ("A", "a"), ("B", "b")]
+        gen = gen_from_choices(choices, nullable=False, blankable=True)
+        for _ in range(100):
+            assert gen() is not None
+
+    def test_includes_blank_when_blankable(self):
+        choices = [("", "empty"), ("A", "a")]
+        gen = gen_from_choices(choices, nullable=True, blankable=True)
+        values = {gen() for _ in range(100)}
+        assert "" in values
+
+    def test_includes_none_when_nullable(self):
+        choices = [(None, "none"), ("A", "a")]
+        gen = gen_from_choices(choices, nullable=True, blankable=True)
+        values = {gen() for _ in range(100)}
+        assert None in values
 
 
 class TestStringFieldsFilling:
@@ -201,6 +228,15 @@ class TestFillingIntFields:
         small_int_field = models.DummyIntModel._meta.get_field("small_int_field")
         assert isinstance(small_int_field, fields.SmallIntegerField)
         assert isinstance(dummy_int_model.small_int_field, int)
+
+    @pytest.mark.skipif(
+        django.VERSION < (5, 0),
+        reason="The db_default field attribute was added after 5.0",
+    )
+    def test_respects_db_default(self):
+        person = baker.make(models.Person, age=10)
+        assert person.age == 10
+        assert person.retirement_age == 20
 
 
 @pytest.mark.django_db
